bifurcating the animation object and constant so we can reuse it across components
