+++ src/App.tsx
@@ -3,6 +3,12 @@ import Button from './shared/components/button/Button';
 
 const App: React.FC = () => {
 
+  const animationObj = {
+    entranceAnimation: 'animate__fadeIn',
+    exitAnimation: 'animate__fadeOut',
+    isEntering: true,
+  };
+
   return (
     <div>
     <Button
@@ -11,6 +17,7 @@ const App: React.FC = () => {
       padding="lg"
       shadow
       onClick={() => console.log("Clicked!")}
+      animationObject={animationObj}
     >
       Submit
     </Button>
