+++ src/shared/components/button/buttonTypes.ts
@@ -1,7 +1,7 @@
 import type { ReactNode } from "react";
 import type { Padding, Margin } from "../../types/spacingTypes";
 import type { SizeValue } from "../../types/sizeTypes";
-import type { AnimationObject } from "../container/containerTypes";
+import type { AnimationObject } from "../../types/animationTypes";
 
 export type ButtonVariant = 'solid' | 'outline' | 'ghost';
 export type ButtonColor = 'primary' | 'secondary' | 'success' | 'error' | 'warning';
