+++ src/shared/types/animationTypes.ts
@@ -1,0 +1,5 @@
+export type AnimationObject = {
+  entranceAnimation: string;
+  exitAnimation: string;
+  isEntering: boolean;
+};
